requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
geopy==2.4.1
streamlit==1.37.1
pytest==8.3.2
openai==1.43.0
langchain-openai==0.1.25
langchain-community==0.2.12
crewai==0.54.1
tavily-python==0.3.6
